init commit

+++ Simulation/controllers/SingleThreadUppaalBot.cpp
@@ -359,7 +359,8 @@ std::vector<int> SingleThreadUppaalBot::getStationPlan(std::string modelOutput) 
 
         //Needed for the initial creation of station plans.
         int tmpLastLocation;
-        if(lastLocation >= sMap.stationIDs.size() + sMap.endStationIDs.size())
+
+        if((long unsigned int)lastLocation >= sMap.stationIDs.size() + sMap.endStationIDs.size())
             tmpLastLocation = sMap.stationIDs.size() + sMap.endStationIDs.size();
         else
             tmpLastLocation = lastLocation;
@@ -573,7 +574,7 @@ void SingleThreadUppaalBot::constructStationUppaalModel(){
     std::string line;
     int numOfStations;
     std::string matrix;
-    if(lastLocation >= sMap.stationIDs.size() + sMap.endStationIDs.size()){
+    if((long unsigned int)lastLocation >= sMap.stationIDs.size() + sMap.endStationIDs.size()){
         matrix = get_expanded_distance_matrix(sMap, lastLocation);
         numOfStations = sMap.stationIDs.size() + sMap.endStationIDs.size() + 1;
     }
@@ -596,7 +597,7 @@ void SingleThreadUppaalBot::constructStationUppaalModel(){
         if(pos != std::string::npos){
             //@todo: Make proper functions to encapsulate the number written!
             // The id matches the last index of the expanded DistMatrix.
-            if(lastLocation >= sMap.stationIDs.size() + sMap.endStationIDs.size()){
+            if((long unsigned int)lastLocation >= sMap.stationIDs.size() + sMap.endStationIDs.size()){
                 line.replace(pos, std::string{"#CUR_STATION#"}.size(),
                              std::to_string(sMap.stationIDs.size() + sMap.endStationIDs.size()));
             }
